+++ changelog.md
@@ -1,3 +1,4 @@
 ## 26.1.2.110
 ### Changes
 - Leaves, Saplings, Leaf Fragments block/item name will be generated automatically (en_us only) if no translation key of the block/item is defined or no translation key for resources type/tree type defined.
+- Custom Tree Type leaves and saplings Block/Item Model will be generated automatically now.